# apps/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "manhwa-backend",
    "description": "Ikiru Bot manhwa scraper API. Backend runs fully on local VPS Postgres. Use Bearer token for protected endpoints. RSS feeds support dual-read params: format (manhwa/manhua/manga) + type (legacy alias), country (KR/CN/JP) + origin (legacy alias).",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/incidents": {
      "get": {
        "summary": "Incidents",
        "operationId": "incidents_api_v1_incidents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/reader/cover": {
      "get": {
        "summary": "Reader Cover",
        "operationId": "reader_cover_api_v1_reader_cover_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/reader/cover-img": {
      "get": {
        "summary": "Reader Cover Public",
        "operationId": "reader_cover_public_api_v1_reader_cover_img_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/reader/proxy": {
      "get": {
        "summary": "Reader Proxy",
        "operationId": "reader_proxy_api_v1_reader_proxy_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/img": {
      "get": {
        "summary": "Img",
        "operationId": "img_api_v1_img_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/cron/status": {
      "get": {
        "summary": "Cron Status",
        "operationId": "cron_status_api_v1_cron_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/cron": {
      "get": {
        "summary": "Cron Trigger",
        "operationId": "cron_trigger_api_v1_cron_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Cron Trigger",
        "operationId": "cron_trigger_api_v1_cron_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "https://scanner.aldifhr.fun",
      "description": "Production (VPS)"
    }
  ]
}